backend/requirements.txt
algoliasearch-django==2.0.0
Django==4.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyYAML==6.0
requests==2.28.1
django-cors-headers==3.13.0
django-environ==0.9.0
numpy==1.23.4
pandas==1.5.1
apyori==1.1.2
psycopg2-binary==2.8.6
gunicorn==20.1.0
spacy==3.4.2
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.1/en_core_web_md-3.4.1-py3-none-any.whl
